+++ src/main/java/com/depromeet/watni/domain/member/domain/Member.java
@@ -1,0 +1,38 @@
+package com.depromeet.watni.domain.member.domain;
+
+import com.depromeet.watni.domain.member.dto.MemberRequestDto;
+import lombok.*;
+
+import javax.persistence.*;
+
+@Table(name = "member")
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Builder(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "member_id")
+    private long id;
+
+    @Column(name = "name")
+    private String name;
+
+    @Column(name = "email", unique = true, nullable = false)
+    private String email;
+
+    @Column(name = "password", nullable = false)
+    private String password;
+
+    public static Member of(MemberRequestDto memberRequestDto) {
+        return Member
+                .builder()
+                .email(memberRequestDto.getEmail())
+                .name(memberRequestDto.getName())
+                .password(memberRequestDto.getPassword())
+                .build();
+    }
+}
