+++ src/main/java/com/depromeet/watni/domain/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.depromeet.watni.domain.member;
+package com.depromeet.watni.domain.member.repository;
 
+import com.depromeet.watni.domain.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
